1) added DB.ExecuteCommand(string sql) to execute raw sql 2) updating tests to reflect schame changes - eg. CustomerID was int, now string, order.EmployeeID used to be order.ProductiD etc

+++ Tests/Test_NUnit/WriteTest.cs
@@ -46,7 +46,7 @@ namespace Test_NUnit
         {
             LinqTestDB db = CreateDB();
 
-            var q = from at in db.Alltypes where at.Int==1 select at;
+            var q = from at in db.Alltypes where at.int_==1 select at;
 
             Alltype row = q.First();
             DbLinq_EnumTest newValue = row.DbLinq_EnumTest == DbLinq_EnumTest.BB
